requirements.txt
asn1crypto==0.24.0
bootstrap-admin==0.4.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
common==0.1.2
cryptography==2.4.2
Django==1.8.5
django-cors-headers==2.4.0
django-crontab==0.7.1
enum34==1.1.6
ffmpeg==1.4
idna==2.7
ipaddress==1.0.22
MySQL-python==1.2.5
Naked==0.1.31
pexpect==4.6.0
Pillow==5.3.0
psycopg2==2.7.6.1
ptyprocess==0.6.0
pycparser==2.19
pycrypto==2.6.1
PyMySQL==0.9.2
pytz==2018.7
PyYAML==3.13
requests==2.20.1
shellescape==3.4.1
simplejson==3.16.0
six==1.11.0
UNKNOWN==0.0.0
urllib3==1.24.1
xlrd==1.1.0
xlwt==1.3.0
